+++ collector/src/main/java/com/bitmovin/analytics/data/SpeedMeasurement.kt
@@ -1,6 +1,7 @@
-package com.bitmovin.analytics.data;
+package com.bitmovin.analytics.data
 
-import java.util.*
+import java.util.Date
+
 
 class SpeedMeasurement() {
 
@@ -8,19 +9,19 @@ class SpeedMeasurement() {
     var duration: Long = 0
 
     // Bytes downloaded
-    var size: Long = 0;
+    var size: Long = 0
 
-    //Time to first byte of this download
-    var timeToFirstByte: Float = 0.0f;
+    // Time to first byte of this download
+    var timeToFirstByte: Float = 0.0f
 
     // Time the download finished
-    lateinit var timestamp: Date;
+    lateinit var timestamp: Date
 
     // HTTP Status of the download measurement
-    var httpStatus: Int = 0;
+    var httpStatus: Int = 0
 
     constructor(duration: Long, size: Long) : this() {
-        this.duration = duration;
-        this.size = size;
+        this.duration = duration
+        this.size = size
     }
 }