+++ collector/src/main/java/com/bitmovin/analytics/adapters/AdAdapter.kt
@@ -6,5 +6,5 @@ interface AdAdapter {
     fun release()
     val isLinearAdActive: Boolean
     val moduleInformation: AdModuleInformation
-    val isAutoplayEnabled: Boolean
+    val isAutoplayEnabled: Boolean?
 }