+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Development
 
 ### Added
+
+- `isTV` to sample to distinguish androidTVs from android mobiles
 - Playback start failure metrics to ExoPlayer
 
 ### Change